Land no-portfolio users on the Portfolio page in its unconfigured state
A first-time / no-portfolio user's magic link defaults to /account/portfolio,
which previously redirected to /account (the dashboard) when no portfolio
existed — so the Portfolio and Dashboard routes looked identical and the
onboarding only ever appeared on the dashboard.

Extract the first-run 'Brief your team' onboarding (EmptyState + LiveTicker)
out of the dashboard route into a shared <Onboarding> component, and render
it in place on /account/portfolio for users with no portfolio yet. Users who
own a portfolio still redirect straight to /portfolios/<slug>; the dashboard
renders the same shared onboarding, so there's no duplication.
