+++ web/app/account/portfolio/page.tsx
@@ -1,17 +1,22 @@
 import { redirect } from "next/navigation";
+import Nav from "@/components/nav";
+import Onboarding from "@/components/account/onboarding";
 import { createSupabaseServerClient } from "@/lib/supabase/server";
 import { getPortfolioForUser } from "@/lib/portfolios-query";
 
 export const dynamic = "force-dynamic";
 
 /**
- * Server-side redirect to the signed-in user's own portfolio detail page
- * (`/portfolios/<slug>`). The nav links here so visitors have a stable
- * "Portfolio" entry that always resolves to whichever slug they own — no
- * need for the nav to fetch the slug client-side.
+ * The signed-in user's "Portfolio" surface. The nav links here so visitors
+ * have a stable entry that always resolves to whichever portfolio they own —
+ * no need for the nav to fetch the slug client-side. It is also the default
+ * landing for a magic-link sign-in (the auth callback's `next`).
  *
- *   * Not signed in → /login?next=/account/portfolio
- *   * Signed in, no portfolio yet → /account (where they'd create one)
+ *   * Not signed in            → /login?next=/account/portfolio
+ *   * Signed in, no portfolio  → render the unconfigured onboarding in place
+ *                                (the "Brief your team" first-run screen), so a
+ *                                first-time user lands on the Portfolio page in
+ *                                its unconfigured state rather than the dashboard
  *   * Signed in with portfolio → /portfolios/<slug>
  */
 export default async function PortfolioRedirect() {
@@ -25,9 +30,33 @@ export default async function PortfolioRedirect() {
   }
 
   const portfolio = await getPortfolioForUser(user.id);
-  if (!portfolio) {
-    redirect("/account");
+  if (portfolio) {
+    redirect(`/portfolios/${portfolio.slug}`);
   }
 
-  redirect(`/portfolios/${portfolio.slug}`);
+  // No portfolio yet — show the Portfolio page in its unconfigured state
+  // (the same first-run onboarding the dashboard uses) instead of bouncing
+  // to /account.
+  let displayName = user.email?.split("@")[0] ?? "there";
+  try {
+    const { data } = await supabase
+      .from("profiles")
+      .select("display_name")
+      .eq("id", user.id)
+      .maybeSingle();
+    if (data?.display_name) displayName = data.display_name;
+  } catch {
+    /* ignore — greeting falls back to the email local-part */
+  }
+
+  return (
+    <>
+      <Nav />
+      <main className="flex-1 w-full">
+        <div className="max-w-[1100px] mx-auto w-full px-4 sm:px-6 py-8 sm:py-10">
+          <Onboarding displayName={displayName} />
+        </div>
+      </main>
+    </>
+  );
 }
